requirements.txt
os-sys
numpy
pandas
glob2
opencv-python
torch
torchvision
pillow
matplotlib
